Add files via upload
One bug in this code. When playing the game, if the user opts to substitute a letter for the hand then refuses to replay the hand, the next hand is printed twice. Could not figure out why but wanted to move on to other coursework. Will hopefully be able to come back and fix this when I develop some more skills.