Add files via upload
This was a very difficult problem set for me. I must have watched a dozen youtube videos on recursion (which I hope I will be greatful for in the long run). The code for this problem set is a slightly modified version of the code presented in this video here: 

https://www.youtube.com/watch?v=4lIQCoG4MnY&list=PLVrUZJ51TIFDkFjF6wp3Z-gJ1sHB6JhVm&index=1

This walkthrough is fantastic and really helped me understand these concepts (along with other vidoes on solving the hanoi towers and basica factorial recursion). I tried to include a lot of comments on the code to help walk myslef through the function when I ineveidbly forgot how I did it in the future.